Implemented welcome email feature for new customers
This commit adds functionality to send a welcome email to customers upon successful creation of their account. When a new customer is created, a welcome email containing important information and instructions is automatically sent to the customer's email address.

Note:The email delivery mechanism has not been configured with SMTP settings yet.Instead of sending emails, the email content is stored in the tmp/mail folder for future reference.